Added support for both PWMs
Initializes Mag and Led PWMs. Mag PWM is constant. LED PWM will
constantly be updated. Currently LED PWM is set to update to the same
period and duty.

Next will be to interface the potentiometer and update the led period
and led duty accordingly.

After that, it's nearly done. I noticed that the Magnet PWM Output is
currently shared with the board LED. Board LED is shared with Boot
Loader Power Sequence. I don't want the Magnet to be powered while the
device is in the boot phase. Will need to look into removing the boot
loader or reroute magnet pwm output to a different pin... Somehow.