Code complete
POT code and PWM code is done.
Locally, I've also loaded the code via a second arduino UNO as an ISP.
This way, the code starts up as soon as power is applied. No more 10s delay
between power on and application start. Therefore, no inadverdant
pulsing of the led which would be powering on the magnet in this
project.